Вынесение логики таймера в отдельный хук
