+++ src/assets/default.scss
@@ -17,5 +17,5 @@ h1, h2, h3, h4, h5, p, span, div {
 }
 
 .text-white {
-	color: $white-color;
-}
+	color: $white-color !important;
+}
