+++ src/components/switcher/Switcher.module.scss
@@ -1,9 +1,8 @@
 @import '@/assets/colors.module.scss';
 
 .layout {
-	width: 100%;
-	height: 80px;
 	background: var(--primary-color);
+	padding: 10px; 
 	justify-content: center;
 	border-radius: 3px;
 	color: $white-color;
