+++ src/pages/main/constants.tsx
@@ -9,7 +9,7 @@ export const Timers = {
     label: 'Pomodoro',
     min: 25,
     max: 60,
-    value: 25,
+    value: 1,
   },
   [TimerKey.timeout]: {
     label: 'Timeout',
